Use start_if_absent.sh to start wallpaper-auto-switcher
Let start_if_absent.sh return non-zero if the target is already
started, in which case autostart will run prepare-wallpaper to set
the wallpaper.

There was a bug in the previous comment that I didn't run
wallpaper-auto-switcher.sh with '&', which blocks the autostart
script.  This will go away in this change.